:children_crossing: ux: Wrong output & STDERR handling
The checker used to ignore STDERR altogether, which might cause WA in
some online judges if the user forgets to remove them before submitting.
This commit changes this behavior by collapsing the inputs whose outputs
matches the answers, this way the correct test cases stay out of view
when debugging. Also, the test cases are colored in red if their STDERR
is non-empty, this way the user can easily be aware of un-removed /
un-commented debug statements before submitting.